Fix some runtime undefined behaviour warnings.
git-svn-id: https://svn.code.sf.net/p/previous/code/branches/branch_filesharing@1553 9ba66e4b-2233-44b8-a46a-7e93b2105c9e

+++ src/sysReg.c
@@ -212,9 +212,9 @@ void SCR_Reset(void) {
         scr_have_dsp_memreset = 1;
         
         if (ConfigureParams.System.nMachineType==NEXT_STATION) {
-            scr1 |= 0xF<<28;
+            scr1 |= (uint32_t)0xF<<28;
         } else {
-            scr1 |= SLOT_ID<<28;
+            scr1 |= (uint32_t)SLOT_ID<<28;
         }
         scr1 |= TYPE_TURBO<<12;
         return;
